refactor(auth): extract shared Tauri deep-link OAuth launcher
fuji and whispering had byte-identical 139-line deep-link launchers in
their #platform/auth seams. Move the logic to a createTauriDeepLinkOAuthLauncher
in @epicenter/auth/oauth-launchers/tauri, parameterized by issuer, clientId,
resource, and redirectUri. Each app's seam collapses to a single launcher
call. The @tauri-apps/* packages are optional peer dependencies on a new
subpath, so web-only consumers of @epicenter/auth never pull them in and the
core stays framework-agnostic.
